Added ik support for custom characters
